feat: Add getFileById method to DBClient class
This commit adds a new method, getFileById, to the DBClient class in the utils/db.js file. This method retrieves a file from the MongoDB collection based on the fileId and userId. It throws an error if the MongoDB client is not connected or if the file is not found. This method is used in the worker.js file for retrieving the file before generating thumbnails.